pyproject.toml
[project]
name = "hyphenation-test"
version = "0.1.1"
description = "..."
authors = [
  { name="Jürgen Schoenemeyer", email="js007@web.de" },
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "certifi>=2024.12.14",
    "dict2xml>=1.7.6",
    "orjson>=3.10.12",
    "pyhyphen>=4.0.4",
    "pyphen>=0.17.0",
    "pyyaml>=6.0.2",
    "result>=0.17.0",
    "xmltodict>=0.14.2",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "types-pyyaml>=6.0.12.20241230",
    "types-xmltodict>=0.14.0.20241009",
]
